Create transmit-sync-down.scpt
You might notice that the actual sync commands in the script are all nested "if's." This is because the actual sync command includes "with delete orphaned items." Thus, the remote appearance will be mirrored on the local. So if you delete remote files, they will be deleted locally. The nested "if's" are there to prevent file deletion in any local folder other than the intended one. While testing this script it is recommended that you only uncomment the line with the "delete orphaned items" once you are thoroughly sure that your folders reliably appear in Transmit's remote and local windows.